+++ HISTORY.rst
@@ -1,7 +1,8 @@
 .. :changelog:
 
+=============
 History
--------
+=============
 
 2.x
 +++++++++++++++++++++++++++++++++++++
@@ -10,4 +11,7 @@ History
 
 * The ``clean`` and `vprepare`` methods of a field can now be defined as a form method with the signature ``clean_fieldname(py_value, **kwargs)`` and ``prepare_fieldname(obj_value, **kwargs)``.
 
+* Added ``prepare`` and ``clean`` methods to the form so a user can overwrite them to store there the logic of pre and post-processing the data, keeping that logic in the form itself instead of in a view.
+
 * Several bugfixes
+